Extract format fn to separate module
