+++ src/index.js
@@ -1,15 +1,5 @@
+import { format } from "./format";
 import { fetchLayout, defaultLayout } from "./layouts";
-
-function format(dirtyStr, dirtyOptions) {
-  const defaultOptions = {
-    layout: defaultLayout,
-  };
-  const text = String(dirtyStr);
-  const options = { ...defaultOptions, ...dirtyOptions };
-
-  const layout = fetchLayout(options.layout);
-  return Array.from(text, (c) => layout.formatKey(c) || c).join("");
-}
 
 function wrapHandlerwithLayout(layout) {
   function validTarget(target) {
